requirements.txt
# --- Core Application & UI ---
streamlit>=1.31.0
pandas>=2.0.0
python-dotenv>=1.0.0

# --- AI, LangChain & Vector Store ---
langchain>=0.1.0
langchain-community>=0.0.10
langchain-groq>=0.0.1
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
groq==0.11.0            # Pinned based on your input (Client for Groq API)

# --- PDF & Report Generation ---
pypdf>=4.0.0
reportlab>=4.0.0

# --- Networking & HTTP ---
requests>=2.31.0
httpx==0.26.0           # Pinned (Dependency for modern async HTTP)
httpcore==1.0.2         # Pinned (Low-level HTTP)
h11==0.14.0
certifi==2023.11.17

# --- Data Validation & Typing ---
pydantic==2.5.3         # Pinned (Core validation library)
pydantic-core==2.14.6
annotated-types==0.6.0
typing_extensions==4.9.0
typing-inspect==0.9.0

# --- Async Utilities ---
anyio==4.1.0
sniffio==1.3.0
exceptgroup==1.1.0

# --- System & Misc Utilities ---
distro==1.9.0
idna==3.6
google-crc32c==1.6.0